[front] chore: Migrate jira mcp server
