+++ front/lib/actions/mcp_internal_actions/constants.ts
@@ -1,9 +1,6 @@
 import type { InternalAllowedIconType } from "@app/components/resources/resources_icons";
 import type { MCPToolStakeLevelType } from "@app/lib/actions/constants";
-import {
-  JIRA_SERVER_INSTRUCTIONS,
-  SALESFORCE_SERVER_INSTRUCTIONS,
-} from "@app/lib/actions/mcp_internal_actions/instructions";
+import { SALESFORCE_SERVER_INSTRUCTIONS } from "@app/lib/actions/mcp_internal_actions/instructions";
 import { INTERACTIVE_CONTENT_INSTRUCTIONS } from "@app/lib/actions/mcp_internal_actions/servers/interactive_content/instructions";
 import { PRODUCTBOARD_SERVER_INSTRUCTIONS } from "@app/lib/actions/mcp_internal_actions/servers/productboard/instructions";
 import { SLIDESHOW_INSTRUCTIONS } from "@app/lib/actions/mcp_internal_actions/servers/slideshow/instructions";
@@ -35,6 +32,7 @@ import { HTTP_CLIENT_SERVER } from "@app/lib/api/actions/servers/http_client/met
 import { HUBSPOT_SERVER } from "@app/lib/api/actions/servers/hubspot/metadata";
 import { IMAGE_GENERATION_SERVER } from "@app/lib/api/actions/servers/image_generation/metadata";
 import { INCLUDE_DATA_SERVER } from "@app/lib/api/actions/servers/include_data/metadata";
+import { JIRA_SERVER } from "@app/lib/api/actions/servers/jira/metadata";
 import { JIT_TESTING_SERVER } from "@app/lib/api/actions/servers/jit_testing/metadata";
 import { MICROSOFT_DRIVE_SERVER } from "@app/lib/api/actions/servers/microsoft_drive/metadata";
 import { MICROSOFT_EXCEL_SERVER } from "@app/lib/api/actions/servers/microsoft_excel/metadata";
@@ -468,48 +466,10 @@ export const INTERNAL_MCP_SERVERS = {
     allowMultipleInstances: true,
     isRestricted: undefined,
     isPreview: false,
-    tools_stakes: {
-      // Read operations - never ask (no side effects)
-      get_issue: "never_ask",
-      get_projects: "never_ask",
-      get_project: "never_ask",
-      get_project_versions: "never_ask",
-      get_transitions: "never_ask",
-      get_issues: "never_ask",
-      get_issues_using_jql: "never_ask",
-      get_issue_types: "never_ask",
-      get_issue_create_fields: "never_ask",
-      get_issue_read_fields: "never_ask",
-      get_connection_info: "never_ask",
-      get_issue_link_types: "never_ask",
-      get_users: "never_ask",
-      get_attachments: "never_ask",
-      read_attachment: "never_ask",
-
-      // Update operations - low stakes
-      create_comment: "low",
-      transition_issue: "low",
-      create_issue: "low",
-      update_issue: "low",
-      create_issue_link: "low",
-      delete_issue_link: "low",
-      upload_attachment: "low",
-    },
     tools_arguments_requiring_approval: undefined,
     tools_retry_policies: undefined,
     timeoutMs: undefined,
-    serverInfo: {
-      name: "jira",
-      version: "1.0.0",
-      description: "Create, update and track project issues.",
-      authorization: {
-        provider: "jira" as const,
-        supported_use_cases: ["platform_actions", "personal_actions"] as const,
-      },
-      icon: "JiraLogo",
-      documentationUrl: "https://docs.dust.tt/docs/jira",
-      instructions: JIRA_SERVER_INSTRUCTIONS,
-    },
+    metadata: JIRA_SERVER,
   },
   interactive_content: {
     id: 23,
